move utilitary classes to new utils folder
