+++ src/main/java/brcomkassin/blockLimiter/utils/ItemBuilder.java
@@ -1,0 +1,52 @@
+package brcomkassin.blockLimiter.utils;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.function.Consumer;
+import java.util.stream.Collectors;
+
+import org.bukkit.ChatColor;
+import org.bukkit.Material;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.ItemMeta;
+
+public final class ItemBuilder {
+
+    private final ItemStack itemStack;
+
+    public ItemBuilder(Material material) {
+        this.itemStack = new ItemStack(material);
+    }
+
+    private ItemBuilder consumeMeta(Consumer<ItemMeta> consumer) {
+        ItemMeta meta = itemStack.getItemMeta();
+        if (meta != null) {
+            consumer.accept(meta);
+            itemStack.setItemMeta(meta);
+        }
+        return this;
+    }
+
+    @SuppressWarnings("deprecation")
+    private String translateColorCodes(String text) {
+        return ChatColor.translateAlternateColorCodes('&', text);
+    }
+
+    @SuppressWarnings("deprecation")
+    public ItemBuilder setName(String displayName) {
+        return consumeMeta(meta -> meta.setDisplayName(translateColorCodes(displayName)));
+    }
+
+    @SuppressWarnings("deprecation")
+    public ItemBuilder setLore(List<String> lines) {
+        return consumeMeta(meta -> meta.setLore(lines.stream().map(this::translateColorCodes).collect(Collectors.toList())));
+    }
+
+    public ItemBuilder setLore(String... lines) {
+        return setLore(Arrays.asList(lines));
+    }
+
+    public ItemStack build() {
+        return this.itemStack;
+    }
+}
